Revert "api,core: interceptor-based config selector (#7610)"
This reverts commit 67bef1457e9e7d629b42fdf1fecf34d443f1cc9b.
