+++ core/src/main/java/io/grpc/internal/ManagedChannelImpl.java
@@ -76,7 +76,6 @@ import io.grpc.internal.ClientCallImpl.ClientStreamProvider;
 import io.grpc.internal.ManagedChannelImplBuilder.FixedPortProvider;
 import io.grpc.internal.ManagedChannelImplBuilder.UnsupportedClientTransportFactoryBuilder;
 import io.grpc.internal.ManagedChannelServiceConfig.MethodInfo;
-import io.grpc.internal.ManagedChannelServiceConfig.ServiceConfigConvertedSelector;
 import io.grpc.internal.RetriableStream.ChannelBufferMeter;
 import io.grpc.internal.RetriableStream.Throttle;
 import java.net.URI;
@@ -898,29 +897,6 @@ final class ManagedChannelImpl extends ManagedChannel implements
     // same target, the new instance must have the same value.
     private final String authority;
 
-    private final Channel clientCallImplChannel = new Channel() {
-      @Override
-      public <RequestT, ResponseT> ClientCall<RequestT, ResponseT> newCall(
-          MethodDescriptor<RequestT, ResponseT> method, CallOptions callOptions) {
-        return new ClientCallImpl<>(
-            method,
-            getCallExecutor(callOptions),
-            callOptions,
-            transportProvider,
-            terminated ? null : transportFactory.getScheduledExecutorService(),
-            channelCallTracer,
-            null)
-            .setFullStreamDecompression(fullStreamDecompression)
-            .setDecompressorRegistry(decompressorRegistry)
-            .setCompressorRegistry(compressorRegistry);
-      }
-
-      @Override
-      public String authority() {
-        return authority;
-      }
-    };
-
     private RealChannel(String authority) {
       this.authority =  checkNotNull(authority, "authority");
     }
@@ -1095,20 +1071,17 @@ final class ManagedChannelImpl extends ManagedChannel implements
 
     private <ReqT, RespT> ClientCall<ReqT, RespT> newClientCall(
         MethodDescriptor<ReqT, RespT> method, CallOptions callOptions) {
-      InternalConfigSelector selector = configSelector.get();
-      if (selector == null) {
-        return clientCallImplChannel.newCall(method, callOptions);
-      }
-      if (selector instanceof ServiceConfigConvertedSelector) {
-        MethodInfo methodInfo =
-            ((ServiceConfigConvertedSelector) selector).config.getMethodConfig(method);
-        if (methodInfo != null) {
-          callOptions = callOptions.withOption(MethodInfo.KEY, methodInfo);
-        }
-        return clientCallImplChannel.newCall(method, callOptions);
-      }
-      return new ConfigSelectingClientCall<>(
-          selector, clientCallImplChannel, executor, method, callOptions);
+      return new ClientCallImpl<>(
+          method,
+          getCallExecutor(callOptions),
+          callOptions,
+          transportProvider,
+          terminated ? null : transportFactory.getScheduledExecutorService(),
+          channelCallTracer,
+          configSelector.get())
+          .setFullStreamDecompression(fullStreamDecompression)
+          .setDecompressorRegistry(decompressorRegistry)
+          .setCompressorRegistry(compressorRegistry);
     }
   }
