+++ core/src/main/java/io/grpc/internal/ManagedChannelServiceConfig.java
@@ -180,7 +180,15 @@ final class ManagedChannelServiceConfig {
     if (serviceMap.isEmpty() && serviceMethodMap.isEmpty() && defaultMethodConfig == null) {
       return null;
     }
-    return new ServiceConfigConvertedSelector(this);
+    return new InternalConfigSelector() {
+      @Override
+      public Result selectConfig(PickSubchannelArgs args) {
+        return Result.newBuilder()
+            .setConfig(ManagedChannelServiceConfig.this)
+            .setCallOptions(args.getCallOptions())
+            .build();
+      }
+    };
   }
 
   @VisibleForTesting
@@ -376,24 +384,6 @@ final class ManagedChannelServiceConfig {
       return new HedgingPolicy(
           maxAttempts, hedgingDelayNanos,
           ServiceConfigUtil.getNonFatalStatusCodesFromHedgingPolicy(hedgingPolicy));
-    }
-  }
-
-  static final class ServiceConfigConvertedSelector extends InternalConfigSelector {
-
-    final ManagedChannelServiceConfig config;
-
-    /** Converts the service config to config selector. */
-    private ServiceConfigConvertedSelector(ManagedChannelServiceConfig config) {
-      this.config = config;
-    }
-
-    @Override
-    public Result selectConfig(PickSubchannelArgs args) {
-      return Result.newBuilder()
-          .setConfig(config)
-          .setCallOptions(args.getCallOptions())
-          .build();
     }
   }
 }
